python_project_1/requirements.txt
requests
beautifulsoup4
pandas
google-genai
lxml